// src/utils/types.d.ts
export interface WeatherInfo {
    country?: string,
    city?: string,
    temp?: number,
    pressure?: number,
    sunset?: number
}

export interface WeatherResponse {
    name: string,
    main: {
        temp: number,
        pressure: number
    },
    sys: {
        country: string,
        sunset: number
    }
}